fix file save and improve config API
The NewConfig takes in a lot of parameters but as the fields in the config
struct increases, it starts to fall short. This replaces NewConfig with a NewCanvas
method that takes in only the width and height. Other fields can then be added/changed
by modifying the returned struct pointer directly.

Also fixes a minor bug where all files are saved with png encoding.